refactor: place 컴포넌트 폴더 내부에 컴포넌트 별로 폴더 생성
