+++ src/app/place/[id]/components/Header/index.tsx
