+++ src/app/place/[id]/components/Info/TimeInfo.tsx
