+++ src/app/place/[id]/components/Info/index.tsx
