+++ src/app/place/[id]/components/index.ts
@@ -1,11 +1,9 @@
-export { default as DefaultInfo } from './DefaultInfo';
 export { default as Header } from './Header';
-export { default as KagongBox } from './KagongBox';
-export { default as ReviewBox } from './ReviewBox';
+export { default as KagongItem } from './KagongItem';
+export { default as ReviewItem } from './ReviewItem';
 export { default as Tag } from './Tag';
-export { default as TimeInfo } from './TimeInfo';
 export { default as Tooltip } from './Tooltip';
+export { default as Info } from './Info';
 export { default as ReviewSheet } from './ReviewSheet';
-export { default as Rating } from './Rating';
 export { default as TimeLogSheet } from './TimeLogSheet';
 export { default as ShareSheet } from './ShareSheet';
