+++ src/app/place/[id]/page.tsx
@@ -1,10 +1,10 @@
 'use client';
 
 import {
-  DefaultInfo,
+  Info,
   Header,
-  KagongBox,
-  ReviewBox,
+  KagongItem,
+  ReviewItem,
   ReviewSheet,
   ShareSheet,
   Tag,
@@ -90,11 +90,11 @@ export default function Page({ params }: { params: { id: string } }) {
         <hr className="my-8 text-bk10" />
 
         <h5 className="mb-4 text-sub1">기본 정보</h5>
-        <DefaultInfo place={placeData} />
+        <Info place={placeData} />
         <h5 className="mb-4 mt-10 text-sub1">카공을 위한 정보</h5>
         <div className="flex w-[calc(100%+1.5rem)] gap-2 overflow-hidden overflow-x-scroll pb-5 pr-6">
           {['CLEAN', 'QUIET', 'SEAT', 'TABLE', 'TEMPERATURE', 'WIFI'].map((type, index) => (
-            <KagongBox key={index} type={type as PlaceConditionType} isFirst={index === 0} />
+            <KagongItem key={index} type={type as PlaceConditionType} isFirst={index === 0} />
           ))}
         </div>
         <a
@@ -142,7 +142,7 @@ export default function Page({ params }: { params: { id: string } }) {
         <hr className="mb-6 text-bk10" />
         <div className="mb-6 flex flex-col gap-5">
           {reviewsData?.pages.map(({ data }) =>
-            data.reviews.map((review) => <ReviewBox key={review.id} review={review} />)
+            data.reviews.map((review) => <ReviewItem key={review.id} review={review} />)
           )}
         </div>
         <Button type="ROUND_DEFAULT" className="mb-10" onClick={() => setIsReviewSheetOpen(true)}>
