+++ src/app/place/[id]/review/page.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { Header, ReviewBox } from '../components';
+import { Header, ReviewItem } from '../components';
 import { useGetReviews } from '@/apis/place';
 import { useRouter } from 'next/navigation';
 
@@ -28,7 +28,7 @@ export default function Page({ params }: { params: { id: string } }) {
       />
       <div className="my-6 flex flex-col gap-5 px-6">
         {reviewsData?.pages.map(({ data }) =>
-          data.reviews.map((review) => <ReviewBox key={review.id} review={review} />)
+          data.reviews.map((review) => <ReviewItem key={review.id} review={review} />)
         )}
       </div>
     </div>
